requirements.txt
streamlit
streamlit-authenticator
langchain
langchain-community
langchain-google-genai
langchain-qdrant
qdrant-client
openai
pypdf
python-dotenv
psycopg2-binary
pyyaml
bcrypt
